+++ twister2/examples/src/java/edu/iu/dsc/tws/examples/basic/BasicKubernetesJob.java
@@ -55,8 +55,7 @@ public final class BasicKubernetesJob {
     int workers = SchedulerContext.workerInstances(config);
     int diskMegaBytes = AuroraContext.workerDisk(config);
     String jobName = SchedulerContext.jobName(config);
-    ResourceContainer resourceContainer =
-        new ResourceContainer((int) cpus, ramMegaBytes, diskMegaBytes);
+    ResourceContainer resourceContainer = new ResourceContainer(cpus, ramMegaBytes, diskMegaBytes);
 
     // build JobConfig
     JobConfig jobConfig = new JobConfig();
